+++ src/Ema/App.hs
@@ -20,7 +20,7 @@ import Data.Some
 import Ema.Asset (Asset (AssetGenerated), Format (Html))
 import Ema.CLI (Cli)
 import Ema.CLI qualified as CLI
-import Ema.Class (Ema(..))
+import Ema.Class (Ema (..))
 import Ema.Generate qualified as Generate
 import Ema.Server qualified as Server
 import System.Directory (getCurrentDirectory)