+++ src/Ema/Example/Ex02_Basic.hs
@@ -1,8 +1,11 @@
+{-# LANGUAGE DeriveAnyClass #-}
+
 -- | A very simple site with two routes, and HTML rendered using Blaze DSL
 module Ema.Example.Ex02_Basic where
 
 import Control.Concurrent (threadDelay)
 import Data.LVar qualified as LVar
+import Data.Universe (Finite, Universe (..), universeGeneric)
 import Ema (Ema (..))
 import Ema qualified
 import Ema.CLI qualified as CLI
@@ -14,7 +17,10 @@ import Text.Blaze.Html5.Attributes qualified as A
 data Route
   = Index
   | About
-  deriving stock (Show, Enum, Bounded)
+  deriving stock (Show, Eq, Generic)
+  deriving anyclass (Finite)
+
+instance Universe Route where universe = universeGeneric
 
 newtype Model = Model {unModel :: Text}